Add small progress dataset to enable visualization
There are several different datasets rather than one because I want to
be able to see how the visualization behaves under several scenarios,
but there are only three scenarios because I don't need more then that
for initial prototypes.

The lack of normalization of the total column isn't a strategic
choice, though it may open up opportunities for visualizing progress
over time even for known out of known projects whose scope of known is
growing over time.

This dataset has not been tested or vetted in any way aside from an
eye check for badly colored syntax. So there might be some syntax
errors. It would be good to setup CI to catch this sort of thing if
the project ever grows beyond a prototype stage, but at this point it
isn't worth the effort investment.