Bootstrap project with create-react-app.
At this stage, there isn't much need for most of what create-react-app
adds, but the scaffolding means I don't have to spend hours working
on scaffolding.

No tests have been done at this point, but we now have the capability
for tests to be done in the future on account of the scaffolding.