Create progress over window function.
_What and Why_

In order to give an estimate of how long things will take we
need an estimate of our progress over time. The simplest
way we can get one in an automated way is to assume that our
past progress over time and our current progress over time
are highly correlated.

Unfortunately, this isn't actually the true rate of progress.

Progress can end up varying. One way we can acknowledge this is by
conceeding that it varies over time and taking several different
samples of progress to get a better estimate of the true progress.
For this reason, we are creating a window function that will let us
get several different probable progress over time metrics.

We need these progress over time metrics, because with them we can
answer questions about possible ways the progress bar might be filled.
We also need it to answer questions about how progress is evolving.

Eventually we're going to use these progress over time metrics to
provide at a glance insight into the way progress is unfolding. Is
progress accelerating as more work is completed? Is progress slowing
down as we near the end of the project? What are the likely end dates
given different rates of progress?

_Testing_

Validated that it gives a reasonable return via eyeballing
the results of a console print, but this function would be
better off being tested with a unit test.