Refactor window creation into its own function.
I want to get tests in place for calculating progress over a window, but the current window method tightly couples the window and the current time. Step one in creating a looser coupling is to get the window creation out of the progressOverWindow function. This sets the stage for doing that by isolating window creation in its own function. In a future diff I'll create the window before calling the progress over window function.

Testing that results didn't change in console.