Refactor progressOverWindow to accept a window.
Moved window creation out of progressOverWindow to decouple windows from calculations related to those windows. Also renamed progressOverWindow to progressDuringWindow; no division of the progress has happened so its misleading to use the word over as a part of the function.

Tested by loading it in the browser and manually checking that the results haven't changed.