Decouple window creation from ThinkProgress component and fixed a bug.
Decoupling the display and the function make it
easier for me to test. Doing this was useful, since it caught a bug: when the first item was the start of the window the prior to the window being accessed was undefined. In that case we should have been using 0 as the initial completed.

Getting the tests working for this function also meant removing some of the React tests since we are no longer displaying what it displayed.

For the first time we have a unit test. Yay!