Refactor and show both table and visualization.
Being able to get at the source of a dataset is important in
data visualization for integrity and credibility reasons. So I
want to have the data available and inspectable. Including it in
a table allows for that.

Refactored, because I view the table and visualization as seperate
components. The table is intended to be a mechanism for getting at
the true data that the visualization is derived from whereas the
visualization is intended to think about what the data has meant and
implies about the future.

Tested by opening browser to observe tha both table and chart are
rendering.