Added data scaling to visualization.
I need the data in the visualization to be sized according to
the visualization context and the data that is being presented.
This introduces a key step in doing that via scaling elements.
At the same time this also introduces support for specifying the
other side of the scaling relationship: the size of the canvas
we will be painting on.

In future steps I'll be modifying this part of the visualization
so as to scale the visualization in proportion to the expected
end date of the task. Right now I'm only doing correct scaling for
the progress dimension, the y axis, and not the time dimension, the x
axis, because I haven't yet calculated the extent of the y axis domain.

I'm also going to need to modify the visualization canvas to support
the inclusion of legends, labels, and other non-data markup which helps
to give context to the data being presented.

This code was tested by rendering in the browser and validating that
a line was drawn.