Create an interpolation function.
Ultimately, the goal is to have the progress visualization show not
just the progress that has been made so far, but likely completion
progressions whose merit is that they are derived from past progress
measures. In order to show these possible completions we need be able
to create them. The interpolation function does just that. It creates
a candidate completion given the current progress, expected end, and
the average progress delta.

In some cases it isn't reasonable to plot completions, because it
would make display unweidly. In recognition of this the interpolation
function has a more complicated output type than merely that of an
array. Right now this isn't encoded specifically via typing and is
instead implicit. Future work should modify the returned interpolation
to be a type that exposes that at times no interpolation is returned
and that the method of displaying the expected progression should use a
more succint information conveyance mechanism.

This was tested via two simple jest tests.