Encapsulate extrapolation of progression.
Extrapolating progress is often going to be wrong, especially in
edge cases like when no progress has been made and so we don't have
data to bootstrap from. So a progression that is always returned as
an array of future progress isn't reasonable if we don't want to deal
with extrapolations of unbound length. Which I don't, because sometimes
the edge case is going to be that only one of the projections is
unreasonable, but all the others are reasonable, and I'd rather not
distort the chart with unreasonable projections in those cases.

All of this means that extrapolation management is a bit more complex
than an array. So we're upgrading it from a returned array, null, and
string to its own type. We're basically lifting our modeling of the
domain up so we can work at a higher level and Encapsulating the
lower level details. Rendering the extraplation doesn't have to wonder
at ideas like 'what is reasonable extrapolation' and can instead
just check to see if an extrapolation should be given and get one if
needed.

Tests adjusted to accomodate the refactoring. Tests are also passsing.