Refactor extrapolation to support returning [x, y]
Ultimately we need to return not just the progression on the
y axis (progress) but also the dates associated with that
progress on the x axis (progress updates). This brings the
structure of the returned response into alignment with our
end goal.

Tests updated and run to reflect change in structure of returned
response.