Return timeline as part of extrapolation.
Timeline is being calculated via d3 datetime libraries because
time can end up being messy and I want to work above that level of
messiness.

We now have the ability to extrapolate progress both in terms of
the timeline and also in terms of the progress made. This isn't
going to be the most accurate with regard to date, because we're
actually just assuming that the progress made on each day is the
progress delta. It could and probably is the case generally that
progress on each day is not uniformly distributed, but for a
prototype this is good enough especially since people should
understand that we're dealing with estimates not prophecy.

Added a new test that verifies that the returned dates are in line
with our expectation. Tests passed.