Refactor to encapsulate analysis object.
We want to analyze the progress over window for several different
windows. The meat of the analysis functions were previously in the
ThinkProgress file, but as the project has evolved this file has
become a place for rendering of tables. As a first step to getting
analysis done, analysis has been lifted up into its own top-level
ThinkProgress class and analysis has been moved from a low-level
visualization of tables to something done in the top-level of the
the ThinkProgress component.

I also adjusted from a function-based declaration to a class-based
declaration so that it is easier to keep track of the various metadata
associated with each analysis.

The mental model for how things work is raw data -> analyse ->
extrapolate from analysis -> communicate the results.

In my next work I'm planning to adjust the analysis object to contain
and calculate relevant metadata. Future work will tie the analsis to
extrapolation. Then from there we can finally proceed to visualizing
the results.

Tested by running npm test.