Create extrapolation objects.
Extrapolations should be done using the analysis objects. Currently,
we don't use the analysis objects directly because it would couple the
analysis object to the extrapolation object such that I would need an
analysis object to analyse an extrapolation.

This isn't tested. It is scaffolding. Future work will fill in the
missing pieces.