Refactor to use getter functions.
Completed, total and endDate are not being set via
getter functions. This paves the way for easier
testing of these fields.

I'm not particularly happy to be special casing
within the analysis object. Really, we shouldn't
even bother analyzing something when we have no
data to analyze. So it might make more sense to
just refuse to do analysis if relevant data is
not provided.

Both testing and visual inspection of the visualization
validated these changes.