Introduces the notion of a ProgressData format.
We need the same format for both progress updates and
extrapolated progress updates. Previously this was
implicit. It still is, but we're moving in the direction
of it being explicit, because we now need to use both
progress data and extrapolated progress data in the
same context.

There are several obvious weaknesses with the way
data is currently represented. Dates being strings
rather than their actual type. The total field being
reproduced. Being implicit rather than explicit and
therefore not checkable via type checking. However,
none of this is being tackled just yet.

Instead we're just trying to conform to the implicit
contract.

Existing tests continue to pass after this change.
An additional test is needed, but not yet added, to
verify that conversion to the not yet formally
specified ProgressData format was successful.