Expand extent with extrapolations.
The extent of our visualization needs to be sized according to
the things we want to render in our visualization. With this,
we accomplish that.

Although tests pass after this change the visualization
hasn't updated to look like what I expected. I'll need
to figure out why in future work.