Refactor Data module name.
This is a very small step in the direction of
being explicit rather than implicit about the
progress dataset format. Instead of calling
things Data we are being explicit about what
type of data we are dealing with. Right now
this is explicitness is only superficial.

Validated via testing and looking at the
visualization.