Add x axis.
Added an X axis to the chart. Right now the x axis isn't
positioned correctly. It is at the top of the chart rather
than the bottom.

I also made a change to the x axis. Before we were using a
linear scale for the x axis, but this is interpreted as a
number not a date when rendered the x axis. The x axis in
our progress visualization is time. So this was incorrect.

Validated visualization by looking at it.