Put visualiation inside a canvas.
If you try to fit everything inside a space for visualization
and your axis and other things are flush with the boundary
of the visualization what can happen and was happening is that
the edges of the visualized axis get caught off.

I'm solving that problem by adding a margin to the visualization
and drawing the visualization within that marging rather than
directly onto the canvas.

I verified visually that ticks and labels are no longer being
cut off.