Extrapolate based on progress per day.
When doing extrapolation the expected progress was all
progress that had been made rather than all the progress
over the window scaled by the size of the window.
This is the same error that was mentioned in a previous
commit with regard to extent sizes. This fixes that
error.

It also adjusts the window lengths, since with the default
lengths I entered originally the window size is larger than
is useful.

Validated visually that extrapolation causes a corresponding
change in the extent.