Add extrapolated line to canvas.
This adds extrapolated lines to the visualization.
Right now this means that a progress line the day
after the last entry extends from that last entry
to the point on the canvas which corresponds with a
hundred percent completion.

Now that we have two different lines on the page one
glaring shortcoming is that we need a way to differentiate
the two different types of progress. That will need to
come in future work.

Verified that extrapolated progress is being rendered
by viewing page.