Change tickFormat for dates on axis.
Month-Day appeals to me as a format, though it has problems.
Really I just want to get away from having it unspecified.
I'm open to changing it to be a written month rather than a
numerical month, but don't want to quibble over that detail
right now.

Tested by looking at results in browsers and reading it to
see that it looks fine.