Ensure reasonable window length choices.
We don't want to consider time before progress
began as part of the estimate for how long we
expect things to take. Filtering down the maximum
window length under analysis to be lower than the
number of days from the start of the progress
tracking until the present moment is one way to
do that.

Tested that results are still the same by
verifying visually.