Refactor dataset.
Time parsing was happening in way more places than it should
have been necessary to have it happen. The problem was I was
using a string when I needed a date. Moved to using a date
and only going to a string when necessary rather than the other
way around.

Tests verified as passing and visual check shows equivalence
with previous work.