Add color scale for progress lines.
Added a color scale so that when progress lines are rendered
the extrapolated lines can be differentiated from the original
lines.

This still needs labeling so that the context of in just what
way the line should be considered different is implicit rather
than implicit.

Verified visually that the lines are different colors.