Start adding a legend.
Now that we have colors we need a legend to help
people identify what the colors mean. This is
just a first crack at doing this and isn't very
appropriate since the text label isn't correct
and the content isn't positioned appropriately.

However, it definitely puts a mispositioned
and badly texted legend onto the screen. This
has been verified visually by looking at the
site.