Refactor visualization to accept ProgressDataset.
Previously the progress entries were passed in rather
than the progress dataset itself. This was updated to
pass in the ProgressDataset object instead. This object
is richer and has information like the name of the
dataset. So it can be used to provide a better legend.

Verified no changes in behavior when viewing visualization
in browser.