Class legend elements.
The way I have things entering right now has a selection that assumes
elements are classed. Since they weren't that could cause problems
in the event of a re-render. The content would be added again and
again because the data would never join successfully. This fixes that
looming future bug.

Verified class was properly set in console elements viewer.