Add no progress entry.
Starting extrapolation from the current day, because we
assume that progress updates are always given. So assuming
progress hasn't been reported despite the presence of the user
on the site is not presenting our extrapolation correctly
to the intended audience.

This isn't intended to be a tool for others to keep track of
progress of someone else, but for someone to have a better
sense for the consequences of thier own progress.

Verified the existence of the lack of movement in the progress
line in browser.