Add table for extrapolated progress.
According to Edward Tufte one of the important rules for
credibility is showing your data. So making it easier for
people to see not just the source of the visualization in
terms of their progress updates but also the raw data of
extrapolated progress is conceptually appropriate.

Verified presence of a very ugly table.