Include starting point in projection.
The extrapolated projection looked off to me because
it assumed that the start of the line shouldn't include
the lack of progress on the current day. It had a skip,
but I think there should be no skipping.

I removed the skip. It is a bit ugly as a solution, because
it is treating the first entry as a special case. I can
fix that later.

Validated that the progress lines look a lot nicer.