Adjust legend positioning.
Made some adjustments to legend positioning and chart
size. My goal is to eventually have a legend which
doesn't get cut off by the end of the visualization.

Right now the way the legend computation is done is kind
of hacky. I just force the size of the canvas to be larger
than position things as necessary. Making these concepts
more clearly communicated in code would help with future
legibility and interpretability.

I also want to spend a bit of time adding mouseover legends
because I'm realizing that I actually want to have many
of these things on one page, but legends end up being
visual clutter in such a design because everything is
going to have similar legends and they will be known
from context.

Validated existence of legible legend.