Render more than one ThinkProgress visualization.
So far we've only dealt with rendering a single progress dataset.
One of the next things we are going to be taking on is rendering
multiple progress datasets. This begins the process of exposing
the new challenges by rendering multiple ThinkProgress visualizations
on the same canvas.

Verified that npm start doesn't complain about compilation errors.