Improve table styling.
Improve table styling by switching to using rounding rather
than resorting to fixed precision calls. I'm still not entirely
satisfied with the look and feel of the table, but it was also
iterated on in this commit and is improved slightly from where
it was before as a consequence.

I don't really see this work as being about making the most
beautiful tables, but the quality of the visualization when
held up against the quality of the table makes the failure of the
table design fairly striking.

I think the best path forward in future work with regard to fixing
the current table design is to leverage existing tools for styling
tables rather than rolling my own. I don't need to reinvent the
table. It already exists. I want to reinvent the progress bar,
because I consider them to be terrible.

Verified that rounding looks right visually.