Upgrade table styling via bootstrap.
My primary goal isn't to re-invent tables on the web.
In the ideal spreadsheets like functionality would
enable a user to play around with the data in the
table in order to see how that might change the
expectations about the future. At this stage though
I'm more concerned with getting things to a point
where I'm happy with how they look; we don't have
data sources yet. It is premature to look at
spreadsheet editing for experimentation.

Verified that the table styling is much improved.