Update README.
Give people who might check out the README a better sense for where
this project is headed and the problem it is aiming to solve.
It is definitely not yet at the point where the problem is solved,
but at least it is far enough along that people passing by might
be able to understand what I'm doing.