Introduce dashboard.
Introducing a new top level collection class for displaying
a collection of ThinkProgress elements. We want this to
encapsulate the behavior of switching between different progress
viewing contexts.

There will eventually be two states. One will be that a progress
element has been selected. In that case only that progress element
will be rendered. The second option is that no progress element
will have been selected. That is the default case that is currently
shown.

Verified visually that there is no perceptual change.