Add skeleton of logic for filtering.
We need the ability to show and hide the progress visualization
based on whether we have selected the visualization. This introduces
two variables in ThinkProgressDashboard which facilitate that. They
are not yet tied into the state of the application since no selection
mechanism has been made.

I've verified that nothing is displayed when a string is supplied as
the filter keyword and that everything is displayed when undefined is
supplied. This filtering works for now, but once IDs are introduced
the filter stage will have an issue. The filter statement needs to
check both that filtering should be done at all and that the ID matches
not just that the ID matches.