Enable selection of progress visualizations via onClick.
Our dashboard now suports clicking on a dashboard to navigate
to a particular progress bar. This lets us drill down from a
(eventually) grid of progress visualizations to the details of
a specific progress visualization.

There are some problems with this navigation paraidgm at the
moment. One of them is that we don't have a way to navigate back.
Another is this SPA design doesn't currently work well in the
context of the web since it doesn't support linking. It also
isn't taking advantage of React's beauty when used in a functional
manner. However, it is progress and I'm hopeful to get to all
the rest.

Verified through manual interaction that links are working
as expected.