Diffentiate between preview and detailed.
I want to fit more information into the top level view of progress
and I'm assuming that if people have looked at a lot of these top
level views that they are already going to have context on what the
visualization means. I'm also planning to allow for further context
by hovering. So it doesn't make sense to include the legend within
the dashboard.

This removes the legend from the dashboard.

Validated visually that the legend is now no longer present when
viewing the dashboard but is still present on details page.