Add text element to preview to give context on dashboard.
This adds a context element to give context to the dashboard.
I think the way I did it was probably not ideal, since it puts
ownership on the preview code. A smarter way to go about this
would be to eliminate references to the preview from the top
level object and instead control the viewport from the dashboard.
However, at least for now, this will do. Future iteration should
take the time to pull previews out of the visualization context.

Tested that there is now a preview link hovering in the corner
of the visualization. It is reminsicent of the preview titles on
blocks org by Mike Bostock.