Refactor think directory location.
Right now the module can't be imported from other apps. It isn't exporting
itself in a way that other React apps can import. I'm working on resolving
this.

Although this doesn't fix that problem it does help me move toard differentiating
the development app that I'm using to test the components and the components I
ultimately want to make available.