Fix empty charts leading to crashes.
When there is no progress data we were crashing during analysis.
Changed analysis to not analyze when there is no data to analyze
and to give a more reasonable result when asking when data collection
began for a dataset that has no entries (crashing is obviously worse
than giving an incorrect start date).

Verified that the chart now renders without crashing when no data is
present.